refactored setting of MMR and status label extracted AccountConverter
